chore: move draw code to ui lib
