+++ libs/ui/include/bulin/ui/imgui/draw.hpp
@@ -1,0 +1,14 @@
+//
+// Created by nickel on 11/30/24.
+//
+#pragma once
+
+#include <bulin/ui/export.hpp>
+
+#include <bulin/application/app.hpp>
+
+namespace bulin
+{
+void BULIN_UI_EXPORT process_key_events(app_context const& ctx, app const& app);
+void BULIN_UI_EXPORT draw(app_context const& ctx, app const& app);
+}  // namespace bulin
